test(executor): split model orchestration and routing decision tests into dedicated files
